# docs/openapi-baseline.json
{"openapi":"3.1.0","info":{"title":"Monikey API","description":"Monikey personal finance backend API.","version":"0.1.0"},"components":{"schemas":{}},"paths":{"/health/live":{"get":{"responses":{"200":{"description":"Default Response"}}}},"/health/ready":{"get":{"responses":{"200":{"description":"Default Response"}}}},"/auth/register":{"post":{"responses":{"200":{"description":"Default Response"}}}},"/auth/login":{"post":{"responses":{"200":{"description":"Default Response"}}}},"/auth/logout":{"post":{"responses":{"200":{"description":"Default Response"}}}},"/auth/me":{"get":{"responses":{"200":{"description":"Default Response"}}}},"/settings":{"get":{"responses":{"200":{"description":"Default Response"}}},"put":{"responses":{"200":{"description":"Default Response"}}}},"/transactions":{"post":{"responses":{"200":{"description":"Default Response"}}},"get":{"parameters":[{"schema":{"type":"string"},"in":"query","name":"cursor","required":false},{"schema":{"type":"integer","minimum":1,"maximum":100,"default":50},"in":"query","name":"limit","required":false},{"schema":{"type":"string","format":"date"},"in":"query","name":"fromDate","required":false},{"schema":{"type":"string","format":"date"},"in":"query","name":"toDate","required":false},{"schema":{"type":"string","enum":["income","expense","transfer"]},"in":"query","name":"type","required":false},{"schema":{"type":"string","format":"uuid"},"in":"query","name":"categoryId","required":false},{"schema":{"type":"string","format":"uuid"},"in":"query","name":"accountId","required":false}],"responses":{"200":{"description":"Default Response"}}}},"/transactions/{id}":{"get":{"parameters":[{"schema":{"type":"string"},"in":"path","name":"id","required":true}],"responses":{"200":{"description":"Default Response"}}}},"/transactions/{id}/reverse":{"post":{"parameters":[{"schema":{"type":"string"},"in":"path","name":"id","required":true}],"responses":{"200":{"description":"Default Response"}}}},"/accounts":{"get":{"responses":{"200":{"description":"Default Response"}}},"post":{"responses":{"200":{"description":"Default Response"}}}},"/credit-cards":{"post":{"responses":{"200":{"description":"Default Response"}}}},"/accounts/{id}":{"patch":{"parameters":[{"schema":{"type":"string"},"in":"path","name":"id","required":true}],"responses":{"200":{"description":"Default Response"}}}},"/accounts/{id}/archive":{"post":{"parameters":[{"schema":{"type":"string"},"in":"path","name":"id","required":true}],"responses":{"200":{"description":"Default Response"}}}},"/bootstrap":{"get":{"responses":{"200":{"description":"Default Response"}}}},"/goals":{"post":{"responses":{"200":{"description":"Default Response"}}}},"/goals/{id}/fund":{"post":{"parameters":[{"schema":{"type":"string"},"in":"path","name":"id","required":true}],"responses":{"200":{"description":"Default Response"}}}},"/categories":{"post":{"responses":{"200":{"description":"Default Response"}}}},"/budgets":{"get":{"responses":{"200":{"description":"Default Response"}}},"post":{"responses":{"200":{"description":"Default Response"}}}},"/budgets/{id}/allocations":{"post":{"parameters":[{"schema":{"type":"string"},"in":"path","name":"id","required":true}],"responses":{"200":{"description":"Default Response"}}}},"/recurring":{"get":{"responses":{"200":{"description":"Default Response"}}},"post":{"responses":{"200":{"description":"Default Response"}}}},"/recurring/{id}/status":{"patch":{"parameters":[{"schema":{"type":"string"},"in":"path","name":"id","required":true}],"responses":{"200":{"description":"Default Response"}}}},"/recurring/{id}/mark-paid":{"post":{"parameters":[{"schema":{"type":"string"},"in":"path","name":"id","required":true}],"responses":{"200":{"description":"Default Response"}}}},"/investments":{"get":{"responses":{"200":{"description":"Default Response"}}}},"/investments/trades":{"post":{"responses":{"200":{"description":"Default Response"}}}},"/investments/dividends":{"post":{"responses":{"200":{"description":"Default Response"}}}},"/receipts":{"post":{"responses":{"200":{"description":"Default Response"}}}},"/receipts/{id}":{"get":{"parameters":[{"schema":{"type":"string"},"in":"path","name":"id","required":true}],"responses":{"200":{"description":"Default Response"}}},"delete":{"parameters":[{"schema":{"type":"string"},"in":"path","name":"id","required":true}],"responses":{"200":{"description":"Default Response"}}}},"/receipts/{id}/process":{"post":{"parameters":[{"schema":{"type":"string"},"in":"path","name":"id","required":true}],"responses":{"200":{"description":"Default Response"}}}},"/receipts/{id}/commit":{"post":{"parameters":[{"schema":{"type":"string"},"in":"path","name":"id","required":true}],"responses":{"200":{"description":"Default Response"}}}},"/reports/summary":{"get":{"responses":{"200":{"description":"Default Response"}}}},"/reports/cash-flow":{"get":{"responses":{"200":{"description":"Default Response"}}}},"/reports/spending-by-category":{"get":{"responses":{"200":{"description":"Default Response"}}}},"/reports/net-worth":{"get":{"responses":{"200":{"description":"Default Response"}}}},"/reports/budget-performance":{"get":{"responses":{"200":{"description":"Default Response"}}}},"/reports/goals":{"get":{"responses":{"200":{"description":"Default Response"}}}},"/reports/investments":{"get":{"responses":{"200":{"description":"Default Response"}}}},"/assistant/messages":{"post":{"responses":{"200":{"description":"Default Response"}}}},"/assistant/categorize-draft":{"post":{"responses":{"200":{"description":"Default Response"}}}},"/imports/batches":{"post":{"responses":{"200":{"description":"Default Response"}}},"get":{"responses":{"200":{"description":"Default Response"}}}},"/imports/batches/{batchId}":{"get":{"parameters":[{"schema":{"type":"string"},"in":"path","name":"batchId","required":true}],"responses":{"200":{"description":"Default Response"}}}},"/imports/batches/{batchId}/transactions":{"post":{"parameters":[{"schema":{"type":"string"},"in":"path","name":"batchId","required":true}],"responses":{"200":{"description":"Default Response"}}},"get":{"parameters":[{"schema":{"type":"string"},"in":"path","name":"batchId","required":true}],"responses":{"200":{"description":"Default Response"}}}},"/imports/batches/{batchId}/commit":{"post":{"parameters":[{"schema":{"type":"string"},"in":"path","name":"batchId","required":true}],"responses":{"200":{"description":"Default Response"}}}},"/imports/plaid/link-token":{"post":{"responses":{"200":{"description":"Default Response"}}}},"/imports/plaid/exchange-token":{"post":{"responses":{"200":{"description":"Default Response"}}}},"/imports/plaid/items":{"get":{"responses":{"200":{"description":"Default Response"}}}},"/imports/plaid/webhook":{"post":{"responses":{"200":{"description":"Default Response"}}}},"/imports/csv/upload":{"post":{"responses":{"200":{"description":"Default Response"}}}},"/openapi.json":{"get":{"responses":{"200":{"description":"Default Response"}}}}},"servers":[{"url":"/api/v1"}]}